+++ src/components/Header/types.ts
@@ -1,0 +1,4 @@
+export interface IFilterTag {
+  href: string;
+  name: string;
+}
