--- src/components/HeaderWarning/style.css
@@ -1,4 +1,0 @@
-.tail-aviso-alternado-lista,
-.tail-aviso-alternado-item{background-color:#000;}
-.tail-aviso-alternado-item,
-.tail-aviso-alternado-link{color:#fff}
